fix fabric drop event is not using correct pair and delay

+++ library/src/main/java/io/callstats/event/fabric/FabricDroppedEvent.kt
@@ -16,7 +16,7 @@ class FabricDroppedEvent(
     val connectionID: String,
     val currIceCandidatePair: IceCandidatePair,
     val prevIceConnectionState: String,
-    val delay: Int) : FabricEvent()
+    val delay: Long) : FabricEvent()
 {
   val currIceConnectionState = "failed"
